requirements.txt
streamlit>=1.34.0
pillow>=10.3.0
pymupdf>=1.24.1
pypdf2>=3.0.1
pandas>=2.2.2
openpyxl>=3.1.2
xlwt>=1.3.0
docx2pdf>=0.1.8
python-docx>=1.1.0
pypandoc>=1.13
streamlit-sortables==0.3.1
chardet==5.2.0